# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /team:
    x-swagger-router-controller: team
    get:
      description: Get team list
      operationId: getTeamList
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TeamListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
    post:
      description: Create new team
      operationId: foundNewTeam
      parameters:
        - in: body
          name: teamdata
          description: team information
          required: true
          schema:
            $ref: "#/definitions/TeamData"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/FoundResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
  /team/dupcheck:
    x-swagger-router-controller: team
    post:
      description: Teamname duplicate check with given teamname
      operationId: teamnameDupCheck
      parameters:
        - in: body
          name: teamname
          description: teamname
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        "201":
          description: Duplicate
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
  /team/{team_id}/members:
    x-swagger-router-controller: team
    get:
      description: Get members of team by using team_id
      operationId: getTeamMemberById
      parameters:
        - in: path
          name: team_id
          description: team_id of team to get members
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
  /team/{team_id}/waiting:
    x-swagger-router-controller: team
    get:
      description: Get waitings of team by using team_id
      operationId: getTeamWaitingById
      parameters:
        - in: path
          name: team_id
          description: team_id of team to get members
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
    post:
      description: Request to join the team with team_id
      operationId:  requestTeamJoin
      parameters:
        - in: path
          name: team_id
          description: team_id of team to get members
          required: true
          type: integer
        - in: body
          name: user_id
          description: user_id of user to join team
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        "400":
          description: Duplicate Request
          schema:
            $ref: "#/definitions/MessageResponse"
        "401":
          description: Already joined
          schema:
            $ref: "#/definitions/MessageResponse"
        "402":
          description: Join Blocked
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
  /team/{team_id}/waiting/management:
    x-swagger-router-controller: team
    post:   
      description: Management waitings of team by using team_id
      operationId: manageTeamWaitingById
      parameters:
        - in: path
          name: team_id
          description: team_id of team to management waitings
          required: true
          type: integer
        - in: body
          name: commitdata
          description: target_user_id and accept or reject to manage waiting
          required: true
          schema:
            $ref: "#/definitions/CommitData"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        "400":
          description: No user with target_user_id waiting to join the team with team_id
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
  /team/{team_id}/recorded:
    x-swagger-router-controller: team
    get:
      description: Get game_id of recorded game list by using team_id
      operationId: getRecordedGameListById
      parameters:
        - in: path
          name: team_id
          description: team_id of team to get recorded game list
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GameListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
...
definitions:
...
  MessageResponse:
    properties:
      message:
        type: string
  TeamListResponse:
    required:
      - team_list
    properties:
      team_list:
        type: array
        items:
          $ref: "#/definitions/TeamOverview"
...
  TeamData:
    required:
      - teamname
      - sector
      - owner_id
      - message
      - option_auto_join
      - option_re_request_join
    properties:
      id:
        type: integer
      teamname:
        type: string
      sector:
        type: string
      owner_id:
        type: integer
      message:
        type: string
      found_dt:
        type: string
        format: date-time
      option_auto_join:
        type: boolean
      option_re_request_join:
        type: boolean
  FoundResponse:
    required:
      - team_id
    properties:
      team_id:
        type: integer
  UserListResponse:
    required:
      - user_list
    properties:
      user_list:
        type: array
        items:
          type: integer
  CommitData:
    required:
      - target_user_id
      - accept_or_reject
    properties:
      target_user_id:
        type: integer
      accept_or_reject:
        type: string
  TeamOverview:
    required:
      - teamname
      - sector
      - owner_id
      - message
      - found_dt
    properties:
      teamname:
        type: string
      sector:
        type: string
      owner_id:
        type: integer
      message:
        type: string
      found_dt:
        type: string
        format: date-time
  GameListResponse:
    required:
      - game_list
    properties:
      game_list:
        type: array
        items:
          $ref: "#/definitions/GameOverviewItem"
  GameOverviewItem:
    required:
      - id
      - hometeam_id
      - awayteam_id
      - game_dt
      - home_score
      - away_score
    properties:
      id:
        type: integer
      hometeam_id:
        type: integer
      awayteam_id:
        type: integer
      game_dt:
        type: string
        format: date-time
      home_score:
        type: integer
      away_score:
        type: integer
...
securityDefinitions:
  api_key:
    type: apiKey
    in: query
    name: api_key
security:
  - api_key: []